requirements.txt
pbr>=0.5.21,<1.0

argparse
kombu
python-jenkins
PyYAML>=3.1.0
Paste
WebOb>=1.2.3,<1.3
paramiko>=1.8.0
GitPython>=0.3.2.RC1
lockfile>=0.8
librabbitmq
ordereddict
oslo.config>=1.2.0
oslo.messaging>=1.3.0a7
python-daemon
extras
statsd>=1.0.0,<3.0
voluptuous>=0.7
gear>=0.4.0,<1.0.0
apscheduler>=2.1.1,<3.0
